make diffusion tests work again
